# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LARAVEL Rest API DOCUMENTATION",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/parking-lot/{id}/slots": {
            "get": {
                "tags": [
                    "Block"
                ],
                "summary": "get all slot in this block with detail status",
                "operationId": "getSlotStatusByBookingDateTime2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the parking lot to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1000000
                    },
                    {
                        "name": "start_datetime",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2023-03-01 14:30:00"
                    },
                    {
                        "name": "end_datetime",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2023-04-01 14:30:00"
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/booking/slots": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "show detail booking",
                "operationId": "getSlotsByIdWithBlockName",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "query",
                        "description": "An array of integers.",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        },
                        "example": "[100000000,100000001]"
                    },
                    {
                        "name": "start_datetime",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2023-03-01 14:30:00"
                    },
                    {
                        "name": "end_datetime",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2023-04-01 14:30:00"
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/booking": {
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "booking now",
                "operationId": "bookParkingLot",
                "parameters": [
                    {
                        "name": "slot_ids[]",
                        "in": "query",
                        "description": "Array of booking IDs",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1000000
                    },
                    {
                        "name": "start_datetime",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2023-01-27 14:50:00"
                    },
                    {
                        "name": "end_datetime",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2023-02-01 14:50:00"
                    },
                    {
                        "name": "licensePlate[]",
                        "in": "query",
                        "description": "Array of booking IDs",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/booking/{userId}/history": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "get history booking of user",
                "operationId": "historyBookingSummary",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "id user booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1000000
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/booking/history/details": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "get history detail booking of user",
                "operationId": "historyBookingDetail",
                "parameters": [
                    {
                        "name": "bookingIds[]",
                        "in": "query",
                        "description": "Array of booking IDs",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/booking/show": {
            "get": {
                "tags": [
                    "QrCode"
                ],
                "summary": "Scan QRcode to get detail booking",
                "operationId": "getDetailQRcode",
                "parameters": [
                    {
                        "name": "bookingIds[]",
                        "in": "query",
                        "description": "Array of booking IDs",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/booking/update": {
            "patch": {
                "tags": [
                    "QrCode"
                ],
                "summary": "Qr Code to confirm complete finish booking",
                "operationId": "completeBooking",
                "parameters": [
                    {
                        "name": "bookingIds[]",
                        "in": "query",
                        "description": "Array of booking IDs",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}